refactor: Refactor validation logic and remove ErrorState
Refactored the validation use cases in the `domain` layer to throw more specific exception types. This change improves error handling by providing more granular details about validation failures for name, email, and password inputs.

Additionally, the `ErrorState` data class has been removed from the `presentation` layer as it is no longer required.
