+++ domain/src/commonMain/kotlin/com/moneyplusplus/domain/usecase/validation/NameValidator.kt
@@ -1,13 +1,13 @@
 package com.moneyplusplus.domain.usecase.validation
 
-import com.moneyplusplus.domain.exception.AppException.ValidationException.Name.Empty
+import com.moneyplusplus.domain.exception.ValidationException.Name.Empty
 
 class NameValidator {
 
     operator fun invoke(name: String) {
         when {
             name.isBlank() ->
-                throw Empty
+                throw Empty()
         }
     }
 }
