+++ domain/src/commonMain/kotlin/com/moneyplusplus/domain/usecase/validation/PasswordValidator.kt
@@ -1,24 +1,18 @@
 package com.moneyplusplus.domain.usecase.validation
 
-import com.moneyplusplus.domain.exception.AppException.ValidationException.Password.Empty
-import com.moneyplusplus.domain.exception.AppException.ValidationException.Password.InvalidPassword
+import com.moneyplusplus.domain.exception.ValidationException.Password.Empty
+import com.moneyplusplus.domain.exception.ValidationException.Password.TooShort
 
 
 class PasswordValidator {
 
     operator fun invoke(password: String) {
-        val hasDigitalNumbers = password.any { it.isDigit() }
-        val hasLetters = password.any { it.isLetter() }
-
         when {
             password.isBlank() ->
-                throw Empty
+                throw Empty()
 
             password.length < 8 ->
-                throw InvalidPassword
-
-            !hasDigitalNumbers || !hasLetters ->
-                throw InvalidPassword
+                throw TooShort()
         }
     }
 }
